Fixes branch

+++ backend/app/features/invitation/service.py
@@ -1,9 +1,12 @@
+from sqlalchemy.exc import IntegrityError
+
 from .repository import user_repo, project_repo, notification_repo
 from app.shared.consts import ResultsCodes
 from app.shared.features.notifications.service import send_notifications_to_client
+import datetime
 
 
-def add_user_in_project(project_name, invited_user_email, owner_id):
+def add_user_in_project(project_id, invited_user_email, owner_id):
     """
     Добавляет пользователя в проект.
 
@@ -13,30 +16,41 @@ def add_user_in_project(project_name, invited_user_email, owner_id):
         owner_id (int): Id автора-владельца проекта
 
     Returns:
+        User: Приглашенный пользователь
         ResultCodes: Результат выполнения
     """
-    project = project_repo.get_by_name(project_name)
+    project = project_repo.get_by_id(project_id)
 
     if project and project.owner_id == owner_id:
         added_user = user_repo.get_by_email(invited_user_email)
         if not added_user:
-            return ResultsCodes.USER_NOT_FOUND
+            return None, ResultsCodes.USER_NOT_FOUND
         if added_user.id == owner_id:
-            return ResultsCodes.CANT_INVITE
+            return None, ResultsCodes.CANT_INVITE
         if added_user and user_repo.user_exists(owner_id):
             if project_repo.is_user_in_project(project.id, added_user.id):
-                return ResultsCodes.USER_IS_IN_ALREADY
-            project_repo.add_user_in_project(project.id, added_user.id)
-            notification_repo.add_notification(project.id, owner_id, added_user.id)
+                return None, ResultsCodes.USER_IS_IN_ALREADY
+            try:
+                project_repo.add_user_in_project(project.id, added_user.id)
+            except IntegrityError as e:
+                print(e)
+                return None, ResultsCodes.USER_IS_IN_ALREADY
+            except Exception as e:
+                print(e)
+                return None, ResultsCodes.UNEXPECTED_ERROR
+
+            notification_repo.add_notification(
+                project.id, owner_id, added_user.id, datetime.datetime.now()
+            )
             send_notifications_to_client(added_user.id)
-            return ResultsCodes.OK
+            return added_user, ResultsCodes.OK
         else:
-            return ResultsCodes.USER_NOT_FOUND
+            return None, ResultsCodes.USER_NOT_FOUND
     else:
-        return ResultsCodes.PROJECT_NOT_FOUND
+        return None, ResultsCodes.PROJECT_NOT_FOUND
 
 
-def delete_user_from_project(project_name, invited_user_email, owner_id):
+def delete_user_from_project(project_id, invited_user_email, owner_id):
     """
     Удаляет пользователя из проекта
 
@@ -48,14 +62,17 @@ def delete_user_from_project(project_name, invited_user_email, owner_id):
     Returns:
         ResultCodes: Результат выполнения
     """
-    project = project_repo.get_by_name(project_name)
+    project = project_repo.get_by_id(project_id)
 
     if project and project.owner_id == owner_id:
         deleted_user = user_repo.get_by_email(invited_user_email)
         if deleted_user and user_repo.user_exists(owner_id):
+            if deleted_user.id == owner_id:
+                return None, ResultsCodes.CANT_DELETE
+
             project_repo.delete_user_from_project(project.id, deleted_user.id)
-            return ResultsCodes.OK
+            return deleted_user, ResultsCodes.OK
         else:
-            return ResultsCodes.USER_NOT_FOUND
+            return None, ResultsCodes.USER_NOT_FOUND
     else:
-        return ResultsCodes.PROJECT_NOT_FOUND
+        return None, ResultsCodes.PROJECT_NOT_FOUND
